chore: rename PlayerTest to PlayerArea, remove commented code and old references
