+++ frontend/src/components/MainContentArea.jsx
@@ -1,7 +1,7 @@
 import ExtraInfo from "./ExtraInfo"
 import MainContentRightBar from "./MainContentRightBar"
 import SessionContentHeader from './singles/SessionContentHeader'
-import PlayerTest from "./PlayerTest"
+import PlayerArea from "./PlayerArea"
 
 const MainContentArea = ({session}) => {
   if (!session) return
@@ -9,7 +9,7 @@ const MainContentArea = ({session}) => {
   return (
     <section className="main-content-area">
       <SessionContentHeader session={session}/>
-      <PlayerTest session={session}/>
+      <PlayerArea session={session}/>
       <ExtraInfo session={session}/>
       <MainContentRightBar session={session}/>
     </section>
