+++ frontend/src/components/PlayerArea.jsx
@@ -1,90 +1,59 @@
-import { useCallback, useEffect, useState } from 'react'
-import { useSearchParams } from 'react-router-dom';
+import { useSearchParams } from "react-router-dom"
+import { useEffect, useState } from "react"
+
 import rrwebPlayer from 'rrweb-player'
-import 'rrweb-player/dist/style.css'
 
 const PlayerArea = ({session}) => {
   const [player, setPlayer] = useState(null)
-  const [events, setEvents] = useState(() => session.events)
+  const [searchParams] = useSearchParams()
 
-  const loadSessionPlayer = useCallback(() => {
-    let replayerDiv = document.querySelector("#replayer")
-    
-    if (replayerDiv?.firstChild) {
-      console.log('player killed')
-      replayerDiv.removeChild(replayerDiv.firstChild)
-    }
-
-
-    try {
-      setPlayer(() => {
-        console.log('new player created')
-        return new rrwebPlayer({
-          target: document.getElementById("replayer"),
-          props: {
-            events: session.events,
-            width: 600,
-            height: 270,
-          },
-      })})
-
-      return player
-    } catch (error) {
-      console.log(error)
-    }
-  }, [events, player, session])
-
-  
   useEffect(() => {
-    if (session.events.length < 1 || player) return 
+    if (session.events.length < 1) return
 
-    loadSessionPlayer()
-  }, [player, events, loadSessionPlayer])
+    (() => {
+
+      let replayerDiv = document.getElementById("replayer")
+
+      if (replayerDiv.firstChild) {
+        replayerDiv.removeChild(replayerDiv.firstChild)
+      }
   
+      try {
+        const newPlayer = 
+          new rrwebPlayer({
+            target: document.getElementById("replayer"),
+            props: {
+              events: session.events,
+              // showController: false,
+              width: 1245,
+              height: 304,
+              // autoPlay: true,
+            },
+          })
+        
+        setPlayer(newPlayer)
+        // return newPlayer
+      } catch (error) {
+        console.log(error)
+      }
 
+    })() 
+  }, [session])
+
+  const playerNavigate = () => {
+    const goToSeconds = searchParams.get('time')
+
+    if (!goToSeconds || !player) return
+
+    player.play(); // prevents session replay from restarting from beginning if replay was at end
+    player.goto(Math.floor(goToSeconds * 1000));
+    player.pause(); // returns to paused state for UX
+  }
+
+  playerNavigate()
 
   return (
-    <div className={`player-area ${session?.id}`}>
       <div id="replayer"></div>
-      <PlayerControls player={player}/>  
-    </div>
-  )
-}
-
-const PlayerControls = ({player}) => {
-  const [searchParams] = useSearchParams();
-  
-  const playerNavigate = async (player) => {
-    let playerState
-    player.addEventListener('ui-update-current-state', (event) => {
-      playerState = event.payload
-    })
-    
-    const seconds = searchParams.get('time')
-
-    if (playerState === "paused") {
-      player.play(); // prevents session replay from restarting from beginning if replay was at end
-      player.goto(Math.floor(seconds * 1000));
-      player.pause(); // returns to paused state for UX
-    } else {
-      player.goto(Math.floor(seconds * 1000));
-    }
-  }
-
-  if (searchParams.get('time') && player) {
-    playerNavigate(player)
-  }
-
-  return (
-    <div className='player-controls'>
-      <button
-        onClick={() => player.play()}
-        >PLAY</button>
-      <button
-        onClick={() => player.pause()}
-      >PAUSE</button>
-
-    </div>
   )
 }
